Remove legacy code, reformat code
