+++ bluetooth/src/main/java/pl/radoslav/bluetooth/BluetoothConnection.kt
@@ -37,7 +37,6 @@ class BluetoothConnection(
                 )
             }
         }.flowOn(Dispatchers.IO)
-
     }
 
     fun write(bytes: ByteArray) {
