+++ bluetooth/src/main/java/pl/radoslav/bluetooth/BluetoothUI.kt
@@ -5,9 +5,7 @@ import android.bluetooth.BluetoothAdapter
 import android.content.Intent
 
 object BluetoothUI {
-    fun startAdvertising(
-        activity: Activity
-    ){
+    fun startAdvertising(activity: Activity) {
         val requestCode = 1
         val discoverableIntent: Intent =
             Intent(BluetoothAdapter.ACTION_REQUEST_DISCOVERABLE).apply {
